#8 - ref: move handleClick func into MessageContext
